# docs/swagger.yaml
basePath: /
definitions:
...
  auth.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        example: bonheur@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      token:
        example: JWT_TOKEN
        type: string
      user:
        $ref: '#/definitions/auth.User'
    type: object
  auth.RegisterRequest:
    properties:
      email:
        example: bonheur@example.com
        type: string
      password:
        example: password123
        minLength: 8
        type: string
      username:
        example: bonheur
        type: string
    required:
    - email
    - password
    - username
    type: object
  auth.RegisterResponse:
    properties:
      user:
        $ref: '#/definitions/auth.User'
    type: object
  auth.User:
    properties:
      email:
        type: string
      id:
        format: int64
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
...
  betting.PlaceBetRequest:
    properties:
      amount:
        type: string
      auto_cashout_multiplier:
        type: string
      bet_number:
        type: integer
      round_id:
        type: integer
    type: object
...
  history.PlayerBet:
    properties:
      amount:
        type: string
      auto_cashout_multiplier:
        type: string
      bet_number:
        type: integer
      can_cancel:
        type: boolean
      can_cashout:
        type: boolean
      cashout_multiplier:
        type: string
      id:
        type: integer
      payout:
        type: string
      potential_payout:
        type: string
      round_id:
        type: integer
      round_number:
        type: integer
      round_status:
        type: string
      status:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:7000
info:
  contact: {}
  description: HTTP API for the Aviator real-money game backend.
  title: Aviator Backend API
  version: "1.0"
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with an email and password and returns a JWT.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Log in
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a PLAYER account. The submitted role is never accepted;
        username and email are required, and the password must be at least 8 characters.
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Register a user
      tags:
      - auth
  /api/bets:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/history.PlayerBet'
            type: array
      security:
      - BearerAuth: []
      summary: List own bets (latest 200)
      tags:
      - betting
    post:
      consumes:
      - application/json
      description: Places one of the user's bets in an open game round.
      parameters:
      - description: Bet details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/betting.PlaceBetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or bet
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Place a bet
      tags:
      - betting
...
  /api/wallet/transactions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: List own wallet transactions (latest 200)
      tags:
      - wallet
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"